+++ src/main/java/mdteam/ait/compat/immersive/TardisPortal.java
@@ -32,16 +32,21 @@ public class TardisPortal extends Portal {
     @Override
     public void tick() {
         super.tick();
-
         if (this.getWorld().isClient()) return;
-        if (this.tardisId == null) return;
+        if (this.tardisId == null) {
+            this.discard();
+            return;
+        }
 
         Tardis found = ServerTardisManager.getInstance().getTardis(this.tardisId);
-        if (found == null) return;
-
+        if (found == null) {
+            AITMod.LOGGER.info("Killing portal (" + this.getId() + ") with tardis (" + this.tardisId + ") as found was null");
+            this.discard();
+            return;
+        }
         if (!found.getDoor().isClosed()) return;
         // we know we are closed and have a tardis so we shouldnt be existing AHHH
-        AITMod.LOGGER.info("Killing portal (" + this.getId() + ") with tardis (" + this.tardisId + ")");
+        AITMod.LOGGER.info("Killing portal (" + this.getId() + ") with tardis (" + this.tardisId + ") as doors are closed");
         this.discard();
     }