+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # 1.2.26
 - Autopilot now only works on lowest speed
 - Flipping the handbrake with autopilot enabled now forces you to land instead of crashing
-- Updated monitor screen
+- Crashing now causes an explosion
 - Monitor can now be rotated freely
 - Added eighth doctor hum
 - Made a different particle show when refueling in a RiftChunk
+- Updated monitor screen
 - Added four new creak sounds
 - Failing to takeoff/land no longer sends a message to the pilot
 - Flight sounds no longer play when out of power